feat(M7.4): add property-based and edge case tests (152 total tests)
- Add property_tests.rs: 16 property-based tests using proptest
  * Polynomial properties: eval homomorphism, witness encoding, zero/constant
  * R1CS properties: valid/invalid witness, constraint linearity
  * Sparse matrix properties: linearity, zero, identity
  * Field arithmetic: addition/multiplication commutativity, modular reduction
  * ZK properties: equivalence, blinding preservation
  * Quotient polynomial existence

- Add edge_cases.rs: 16 boundary condition tests
  * Minimal configurations: m=1, m=2 constraints
  * Boundary values: zero, one, q-1 (max field element)
  * Large witness sizes: n=64, n=128
  * Polynomial edge cases: degree 0, 1, high degree
  * Deterministic proof generation with seeds
  * Empty/single-entry sparse matrices

- Fix u64 overflow in modular arithmetic:
  * polynomial.rs: Horner's method uses u128 for addition
  * property_tests.rs: all modular additions wrapped in u128
  * Prevents panic on large field values (≈ q-1)

Progress: 116 baseline → 152 tests (+36, target 200+)

+++ rust-api/lambda-snark/src/polynomial.rs
@@ -103,7 +103,7 @@ impl Polynomial {
         
         for coeff in self.coeffs.iter().rev().skip(1) {
             // result = result * alpha + coeff (mod q)
-            result = (mul_mod(result, alpha.value(), self.modulus) + coeff.value()) % self.modulus;
+            result = ((mul_mod(result, alpha.value(), self.modulus) as u128 + coeff.value() as u128) % self.modulus as u128) as u64;
         }
         
         Field::new(result)
